Handle DABL Login route. cleanup code
