+++ ui/src/components/App.js
@@ -1,17 +1,17 @@
-import React from "react";
+import React, {useEffect} from "react";
 import { HashRouter, Route, Switch, Redirect } from "react-router-dom";
 import Layout from "./Layout/Layout";
 import Error from "../pages/error/Error";
 import Login from "../pages/login/Login";
-import { useUserState } from "../context/UserContext";
+import { useUserState, useUserDispatch } from "../context/UserContext";
 
 export default function App() {
   const userState = useUserState();
-  
+
   return (
     <HashRouter>
       <Switch>
-        <Route exact path="/" render={() => <Redirect to="/app/report" />} />
+        <Route exact path="/" component={RootRoute} />
         <Route
           exact
           path="/app"
@@ -25,6 +25,30 @@ export default function App() {
   );
 
   // #######################################################################
+
+  function RootRoute(props) {
+    var userDispatch = useUserDispatch();
+
+    useEffect(() => {
+      const url = new URL(window.location.toString());
+      const token = url.searchParams.get('token');
+      if (token === null) {
+        return;
+      }
+      const party = url.searchParams.get('party');
+      if (party === null) {
+        throw Error("When 'token' is passed via URL, 'party' must be passed too.");
+      }
+      localStorage.setItem("daml.party", party);
+      localStorage.setItem("daml.token", token);
+
+      userDispatch({ type: "LOGIN_SUCCESS", token, party });
+    })
+
+    return (
+      <Redirect to="/app/report" />
+    )
+  }
 
   function PrivateRoute({ component, ...rest }) {
     return (
